created an env_manager
