+++ src/builtins/unset.c
@@ -6,7 +6,7 @@
 /*   By: malde-ch <malo@chato.fr>                   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/19 14:49:04 by malde-ch          #+#    #+#             */
-/*   Updated: 2025/02/19 17:07:21 by malde-ch         ###   ########.fr       */
+/*   Updated: 2025/02/19 21:33:26 by malde-ch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
